refactor: split perfom auth check into separate hook
